// components/Web4TSComponent/0.3.0.10/test/data/TestCreateComponent/0.1.0.0/src/ts/layer3/Scenario.interface.ts
/**
 * SPDX-License-Identifier: AGPL-3.0-only WITH AI-GPL-Addendum
 * SPDX-FileComment: See ../../../../../../../../../../AI-GPL.md for AI-specific terms.
 * Copyright (c) 2025 Cerulean Circle GmbH
 * Copyleft: See AGPLv3 (../../../../../../../../../../LICENSE) and AI-GPL Addendum (../../../../../../../../../../AI-GPL.md)
 * Backlinks: /LICENSE, /AI-GPL.md
 */

export interface Scenario<T = any> {
  ior: {
    uuid: string;
    component: string;
    version: string;
  };
  owner: string;
  model: T;
}